voice-translator/requirements.txt
discord.py-self==2.1.0
faster-whisper==0.10.0
edge-tts==6.1.12
aiohttp==3.9.1
python-dotenv==1.0.0
numpy==1.26.2
soundfile==0.12.1
PyAudio==0.2.14
